requirements.txt
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-mysql==3.11.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==3.4.0
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
numpy==1.20.1
packaging==20.9
pandas==1.2.2
Pillow==8.1.0
Pint==0.16.1
Pygments==2.8.0
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
uWSGI==2.0.18
zipp==3.4.0